Restrict perms on directory containing Unix socket.
Change the permissions on the directory holding the unix socket by
forking chmod (yes, yes, hackish). There doesn't seem to be a way of
setting perms/mask natively in Erlang.

The alternative is doing the chmod within the NIF (e.g., procket:chmod/2).
At that point, it's maybe better just to make an NIF procket:tmpdir/0.