Find interface MAC address and ARP cache lookup
Add functions to return the MAC address of a device or lookup an IP
address in the ARP cache. Since I couldn't get the ioctl(SIOCGARP)
version to work, packet simply parses the output of /proc/net/arp.
packet:arplookup(Socket, Device, IP) is sort of hacky; it throws away
the first 2 arguments. I left them in since they'll probably be needed
for the BSD version.