Retrieve a list of interfaces on Linux
Retrieve a list of interfaces by parsing the output of /proc/net/dev.
The device list can also be retrieved by using ioctl(SIOCGIFCONF), but
this requires passing in a struct with a pointer to a buffer. Possibly
there could be an NIF allocator that returns the integer pointing to
the memory and a corresponding free().