Allocate pointers to buffers for ioctl requests
Some ioctl requests require a pointer to a user allocated buffer. For
example:

struct  iw_point
{
    void *pointer;
    u_int16_t length;
    u_int16_t flags;
};

Would be constructed using:

{ok, Structure, [Resource]} = procket:alloc([
    {ptr, 4096},                                % buffer
    <<4096:2/unsigned-native-integer-unit:8,    % length
    0:2/native-unsigned-integer-unit:8>>        % flags
]).

A tuple containing the structure and a list of resources (one for
each allocated buffer) is returned.

After calling the ioctl, the contents of the buffer is retrieved by
using the buf/1 function on the resource.

NOTE: If the pointer points to somewhere invalid, ioctl will return
{error, efault}, but if the pointer points somewhere within beam's memory
space, the VM may crash.