Allow setuid open of any character device
Make the bpf device code generic so that the spawned setuid procket
executable is able to open any character device.

The code first checks the multiplex device (e.g., /dev/bpf). If it exists
and no error occurs, the fd is returned to the Erlang side.

If the muliplexing device does not exist or is not a character device,
the code will fall back to trying numbered devices.