Return POSIX atom on error from forked setuid bin
Add a function to the NIF returning an atom from erl_errno_id().

Do not distinguish between procket_cmd and procket errors. Return the
same tuple {error, posix()} on errors.

Return the POSIX value in hackish way: by echo'ing the return code
in os:cmd/1. Otherwise, sudo may return failure (for example, if the
user has not placed the procket command in sudoers.conf. sudo exits
with a return value of EPERM in this case).