Ensure binaries are zero'ed
The binaries returned by some functions are grown to indicate to the
caller the size required, e.g., recvfrom/2 will return the number of
bytes needed for the address parameter if the provided address structure
is too small.

enif_realloc_binary() does not zero allocated bytes, so internal VM data
may be leaked out to the caller.

Create a macro to resize binaries that zero's any trailing data.