Support IPv6 sockets
Add support for IPv6 sockets. By default, the protocol family PF_UNSPEC
is used. Which family is chosen (PF_INET or PF_INET6) depends on the
configuration of the OS (for example, on Ubuntu, this can be set in
/etc/gai.conf). The first protocol family that returns a socket is used.

To explicitly use a a family, pass the {family, Family} option to open/2.
For IPv4, the option is inet; for IPv6, use inet6.