ioctl: allow integer values for argp
The third argument to ioctl is usually a pointer but the FreeBSD man
page clarifies:

    Most uses of ioctl(), however, require the third argument to be a
    caddr_t or an int.

(The size and signedness of the int is not specified.)

For example, tun/tap devices on Linux set persist mode using the integer
value 0 or non-zero.

Unfortunately, allowing integer values allows the caller to mistakenly
pass in an integer where a pointer is required. While this will probably
result in EFAULT or EINVAL, there's a chance the memory of the process
could be corrupted.

Whitelisting ioctl requests with integer values wouldn't be practical
since procket would need to know about every driver.