libancillary: fix alignment of control message
Transferring fd's between processes was failing. Disabling compiler
optimizations resulted in a segfault when ancil_send_fd was called and
on one occasion, a kernel panic (after playing with the cmsg lengths).

Align the buffer using a union as suggested in UNP. Don't use the
libancillary structure for holding the fd's; use a union allocated on the
stack and calculate lengths using the macros. This change only supports
passing a single fd between processes so it isn't a proper fix.