Fix the echo example
The behaviour of gen_tcp:listen/2 seems to have changed at some point:
if a non-zero port is passed in, the inet driver attempts to bind this
port, even if an 'fd' parameter exists in the options.

Setting the port to 0 seems to force inet to use the fd option.