Fix compilation on Solaris
So errno is not thread local on Solaris and is set to 0 by default on
error. Really Solaris?

On Solaris, compile using POSIX mode which requires c99 mode which fixes
errno, uses void* instead of c_addr_t* for iovec and various other errors.

Add solaris constants for dgram, stream and raw which differ from other
unix platforms. Make the send/recvmsg example work on Solaris.

Platform dependent constants need to be fixed in the future. This could
be done either by adding a lookup table to the NIF or writing a small
program in C to generate an erlang module at compile time.

Bump version to 0.6.1.