Allow atoms for socket option parameters
Allow the use of atoms representing the names for socket option level
and option_name parameters. This provides a way for applications using
the procket libary that need to manipulate or inspect socket options to
work across different operating systems without a lot of brittle
application-level checking. These changes affect the setsocketopt and
getsockopt functions and the previous behavior of accepting integers for
the level and option_name parameters is still intact.