procket_cmd: bind ICMP sockets
Allow setting the source address of ICMP sockets by calling bind(2) in
the setuid helper. See:

https://github.com/msantos/gen_icmp/issues/15

For SOCK_RAW sockets, getaddrinfo(3) on Linux expects:

* the service name (aka port) to be NULL
* the node name to be non-NULL

According to the Linux and FreeBSD man page, if the node name is NULL,
default values of INADDR_ANY for IPv4 and IN6ADDR_ANY_INIT for IPv6 will
be used. On Linux at least, passing NULL will result in an error. Set
defaults of either "0.0.0.0" or "::" to avoid the error. Portability
of these defaults still needs testing.

Alternatively, setting of these options could be moved to erlang code,
either in procket.erl or by the caller.