nit: recvfrom/2: ok: match any term
The recvfrom NIF will always return an empty binary if the struct socket
address length is set to 0. If the NIF did return a non-empty binary
(through a bug or code change) then a 3 element ok tuple would be
returned to the caller.

Match any term for the binary to follow the type spec.

Ref: 92c0bde92